added: get request for weather api

+++ app/src/main/java/com/baris/weatherapp/domain/model/Wind.kt
@@ -1,6 +1,6 @@
 package com.baris.weatherapp.domain.model
 
 data class Wind(
-    val deg: Int,
-    val speed: Int
+    val deg: Double,
+    val speed: Double
 )
